requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Measurement tools
codecarbon>=2.3.4
psutil>=5.9.8
py-cpuinfo>=9.0.0

# Data processing
pyyaml>=6.0
jsonschema>=4.17.0

# Analysis & visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0


#LLMS
# Aggiungi queste righe
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
dashscope>=1.14.0  # Alibaba Qwen
together>=1.0.0    # Per Llama 4 via Together.ai
tiktoken>=0.5.0    # Token counting OpenAI
